+++ MoPubSDK/Internal/Banners/MPBannerAdManager.h
@@ -23,10 +23,4 @@
 - (void)startAutomaticallyRefreshingContents;
 - (void)rotateToOrientation:(UIInterfaceOrientation)orientation;
 - (void)cancelAd;
-// Deprecated.
-- (void)customEventDidLoadAd;
-- (void)customEventDidFailToLoadAd;
-- (void)customEventActionWillBegin;
-- (void)customEventActionDidEnd;
-
 @end