+++ MoPubSDK/Internal/Banners/MPBannerAdManager.m
@@ -13,7 +13,6 @@
 #import "MPError.h"
 #import "MPTimer.h"
 #import "MPConstants.h"
-#import "MPLegacyBannerCustomEventAdapter.h"
 
 #import "WBAdEvent_Internal.h"
 #import "WBAdControllerEvent.h"
@@ -398,61 +397,10 @@
     }
 }
 
-#pragma mark - Deprecated Public Interface
-
-- (void)customEventDidLoadAd
-{
-    if (![self.requestingAdapter isKindOfClass:[MPLegacyBannerCustomEventAdapter class]]) {
         AdLogType(WBAdLogLevelWarn, self.logType, @"-customEventDidLoadAd should not be called unless a custom event is in "
-                  @"progress.");
-        return;
-    }
-
-    //NOTE: this will immediately deallocate the onscreen adapter, even if there is a modal onscreen.
-
-    [self.onscreenAdapter unregisterDelegate];
-    self.onscreenAdapter = self.requestingAdapter;
-    self.requestingAdapter = nil;
-
-    [self.onscreenAdapter didDisplayAd];
-
-    [self scheduleRefreshTimer];
-}
-
-- (void)customEventDidFailToLoadAd
-{
-    if (![self.requestingAdapter isKindOfClass:[MPLegacyBannerCustomEventAdapter class]]) {
         AdLogType(WBAdLogLevelWarn, self.logType, @"-customEventDidFailToLoadAd should not be called unless a custom event is in "
-                  @"progress.");
-        return;
-    }
-
-    [self loadAdWithURL:self.requestingConfiguration.failoverURL];
-}
-
-- (void)customEventActionWillBegin
-{
-    if (![self.onscreenAdapter isKindOfClass:[MPLegacyBannerCustomEventAdapter class]]) {
         AdLogType(WBAdLogLevelWarn, self.logType, @"-customEventActionWillBegin should not be called unless a custom event is in "
-                  @"progress.");
-        return;
-    }
-
-    [self.onscreenAdapter trackClick];
-    [self userActionWillBeginForAdapter:self.onscreenAdapter];
-}
-
-- (void)customEventActionDidEnd
-{
-    if (![self.onscreenAdapter isKindOfClass:[MPLegacyBannerCustomEventAdapter class]]) {
         AdLogType(WBAdLogLevelWarn, self.logType, @"-customEventActionDidEnd should not be called unless a custom event is in "
-                  @"progress.");
-        return;
-    }
-
-    [self userActionDidFinishForAdapter:self.onscreenAdapter];
-}
-
 @end