+++ MoPubSDK/Internal/Interstitials/MPInterstitialAdManager.h
@@ -27,9 +27,4 @@
                              testing:(BOOL)testing;
 - (void)presentInterstitialFromViewController:(UIViewController *)controller;
 
-// Deprecated
-- (void)customEventDidLoadAd;
-- (void)customEventDidFailToLoadAd;
-- (void)customEventActionWillBegin;
-
 @end