+++ MoPubSDK/Native Ads/Internal/MPMoPubNativeAdAdapter.h
@@ -8,8 +8,9 @@
 
 @interface MPMoPubNativeAdAdapter : NSObject <MPNativeAdAdapter>
 
-@property (nonatomic, strong) NSArray *impressionTrackers;
-@property (nonatomic, strong) NSURL *engagementTrackingURL;
+@property (nonatomic, weak) id<MPNativeAdAdapterDelegate> delegate;
+@property (nonatomic, readonly) NSArray *impressionTrackerURLs;
+@property (nonatomic, readonly) NSArray *clickTrackerURLs;
 
 - (instancetype)initWithAdProperties:(NSMutableDictionary *)properties;