+++ MoPubSampleApp/Controllers/MPAdTableViewController.m
@@ -17,19 +17,11 @@
 #import "MPGlobal.h"
 #import "MPAdPersistenceManager.h"
 #import "MPAdEntryViewController.h"
-#import "MPNativeAdTableViewController.h"
 #import "MPNativeAdPlacerTableViewController.h"
 #import "MPNativeAdPlacerCollectionViewController.h"
 #import "MPNativeAdPlacerPageViewController.h"
 #import "MPSampleAppLogReader.h"
-
-typedef enum
-{
-    MPAdTableSection_Banner,
-    MPAdTableSection_Interstitital,
-    MPAdTableSection_Native,
-    MPAdTableSection_Saved
-} MPAdTableSection;
+#import "MPRewardedVideoAdDetailViewController.h"
 
 @interface MPAdTableViewController () <UIAlertViewDelegate, UIActionSheetDelegate>
 
@@ -83,6 +75,8 @@ typedef enum
     UIButton* myInfoButton = [UIButton buttonWithType:UIButtonTypeInfoLight];
     [myInfoButton addTarget:self action:@selector(infoButtonClicked:) forControlEvents:UIControlEventTouchUpInside];
     self.navigationItem.leftBarButtonItem = [[UIBarButtonItem alloc] initWithCustomView:myInfoButton];
+
+    [[MPSampleAppLogReader sharedLogReader] beginReadingLogMessages];
 
     [[MPSampleAppLogReader sharedLogReader] beginReadingLogMessages];
 
@@ -143,7 +137,8 @@ typedef enum
     cell.textLabel.textColor = [UIColor colorWithRed:0.42 green:0.66 blue:0.85 alpha:1];
     cell.detailTextLabel.textColor = [UIColor colorWithRed:0.86 green:0.86 blue:0.86 alpha:1];
 
-    cell.accessoryType = indexPath.section == MPAdTableSection_Saved ? UITableViewCellAccessoryDetailDisclosureButton : UITableViewCellAccessoryNone;
+    MPAdSection *section = self.sections[indexPath.section];
+    cell.accessoryType = [section.title isEqualToString:@"Saved Ads"] ? UITableViewCellAccessoryDetailDisclosureButton : UITableViewCellAccessoryNone;
 
     return cell;
 }
@@ -173,11 +168,11 @@ typedef enum
         case MPAdInfoInterstitial:
             detailViewController = [[MPInterstitialAdDetailViewController alloc] initWithAdInfo:info];
             break;
+        case MPAdInfoRewardedVideo:
+            detailViewController = [[MPRewardedVideoAdDetailViewController alloc] initWithAdInfo:info];
+            break;
         case MPAdInfoNative:
             detailViewController = [[MPNativeAdDetailViewController alloc] initWithAdInfo:info];
-            break;
-        case MPAdInfoNativeInTableView:
-            detailViewController = [[MPNativeAdTableViewController alloc] initWithAdInfo:info];
             break;
         case MPAdInfoNativeInCollectionView:
             detailViewController = [[MPNativeAdPlacerCollectionViewController alloc] initWithAdInfo:info];